[!!!][TASK] Make usage of compatibility6 classes configurable
A little history:
TYPO3 removed the inline translation of flexforms in version 7.6.
When I realised that I talked to a core team member and I was told
to integrated the changes of the extension compatibility6 into
templavoila. That however, was quite a huge mistake as compatility6
is not really stable and breaks a lot of flexform handling in 7.6.

As it is simply not possible to fulfil the needs of both worlds
with just one codebase, I decided to make the choice – whether to
use compatibility6 classes or not – to the user.

The fix:
This commit introduces an extension configuration option to enable
the loading of compatibility6 classes. Therefore, this commit is
breaking. By default templavoila does now rely on the regular 7.6
behaviour of parsing interpreting flexforms.

If you are in need of translating your content in flexforms, rather
than having regular overlays of your records (this is the default!)
then, and only then, should you consider enabling the compatiblity6
mode.